# docs/generated/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "lab API",
    "description": "Homelab service orchestration API",
    "contact": {
      "name": "jmagar"
    },
    "license": {
      "name": "MIT OR Apache-2.0",
      "identifier": "MIT OR Apache-2.0"
    },
    "version": "0.3.2"
  },
  "paths": {
...
    "/v1/doctor": {
      "post": {
        "tags": [
          "doctor"
        ],
        "summary": "Dispatch action to doctor",
        "description": "Execute an action on the doctor service. Use `action: \"help\"` to list available actions.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful action response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad request (unknown action, confirmation required)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorUnknownAction"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorSdk"
                }
              }
            }
          },
          "422": {
            "description": "Validation error (missing or invalid param)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMissingParam"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/v1/extract": {
      "post": {
        "tags": [
          "extract"
        ],
        "summary": "Dispatch action to extract",
        "description": "Execute an action on the extract service. Use `action: \"help\"` to list available actions.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful action response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad request (unknown action, confirmation required)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorUnknownAction"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorSdk"
                }
              }
            }
          },
          "422": {
            "description": "Validation error (missing or invalid param)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMissingParam"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/v1/marketplace": {
      "post": {
        "tags": [
          "marketplace"
        ],
        "summary": "Dispatch action to marketplace",
        "description": "Execute an action on the marketplace service. Use `action: \"help\"` to list available actions.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful action response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad request (unknown action, confirmation required)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorUnknownAction"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorSdk"
                }
              }
            }
          },
          "422": {
            "description": "Validation error (missing or invalid param)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMissingParam"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
    "/v1/setup": {
      "post": {
        "tags": [
          "setup"
        ],
        "summary": "Dispatch action to setup",
        "description": "Execute an action on the setup service. Use `action: \"help\"` to list available actions.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ActionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful action response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          },
          "400": {
            "description": "Bad request (unknown action, confirmation required)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorUnknownAction"
                }
              }
            }
          },
          "401": {
            "description": "Authentication failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorSdk"
                }
              }
            }
          },
          "422": {
            "description": "Validation error (missing or invalid param)",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorMissingParam"
                }
              }
            }
          }
        },
        "security": [
          {
            "bearer_auth": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ActionRequest": {
        "type": "object",
        "description": "Shared request type for all service action dispatchers.\n\nEvery service handler deserializes `POST /v1/<service>` bodies into this\nstruct and forwards `action` + `params` to the corresponding dispatch\nfunction, keeping HTTP and MCP input semantics aligned while each transport\nowns its response envelope.",
        "required": [
          "action"
        ],
        "properties": {
          "action": {
            "type": "string"
          },
          "params": {}
        }
      },
...
      "ErrorMissingParam": {
        "type": "object",
        "description": "Error envelope for `missing_param` responses.",
        "required": [
          "kind",
          "message",
          "param"
        ],
        "properties": {
          "kind": {
            "type": "string",
            "description": "Always `\"missing_param\"`."
          },
          "message": {
            "type": "string",
            "description": "Human-readable message."
          },
          "param": {
            "type": "string",
            "description": "The missing parameter name."
          }
        }
      },
      "ErrorSdk": {
        "type": "object",
        "description": "Error envelope for SDK pass-through errors (`auth_failed`, `rate_limited`, etc.).",
        "required": [
          "kind",
          "message"
        ],
        "properties": {
          "kind": {
            "type": "string",
            "description": "Stable kind tag from the SDK (e.g. `\"auth_failed\"`, `\"rate_limited\"`)."
          },
          "message": {
            "type": "string",
            "description": "Human-readable message."
          }
        }
      },
      "ErrorUnknownAction": {
        "type": "object",
        "description": "Error envelope for `unknown_action` responses.",
        "required": [
          "kind",
          "message",
          "valid"
        ],
        "properties": {
          "hint": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional fuzzy match suggestion."
          },
          "kind": {
            "type": "string",
            "description": "Always `\"unknown_action\"`."
          },
          "message": {
            "type": "string",
            "description": "Human-readable message."
          },
          "valid": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Valid action names for this service."
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "token"
      }
    }
  }
}